AgentAsCode_BE/project/18a384a0-bf52-4963-adfb-9068c0cbd6d1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.0
pydantic==2.5.0
google-adk==1.14.1
google-generativeai==0.3.2
pandas
pycryptodome
